Dont run server tests on windows
