+++ .github/workflows/test.yml
@@ -25,6 +25,7 @@ jobs:
     strategy:
       matrix:
         platform: [ubuntu-latest, macos-latest, windows-latest]
+        test-server: [true, true, false]
     runs-on: ${{ matrix.platform }}
     steps:
       - uses: actions/checkout@v3
@@ -32,5 +33,5 @@ jobs:
         with:
           node-version: "16"
       - run: npm install
-      - run: tests/test.sh
+      - run: TEST_SERVER={{matrix.test-server}} tests/test.sh
         shell: bash
